requirements.txt
fastapi
uvicorn[standard]
qdrant-client[fastembed]
openai>=1.0.0
python-dotenv
httpx
pydantic
PyPDF2
